% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratopt.R
\name{BarCy}
\alias{BarCy}
\title{BarCy
Crée et affiche des graphiques basés sur le type de vue sélectionné.}
\usage{
BarCy(
  viewType,
  colorsM,
  colorsR,
  proportions,
  totalTonnageParClasse,
  totalTonnageParRegion,
  totalTonnageParMaison,
  sitesOutreMer
)
}
\arguments{
\item{viewType}{Le type de vue à générer, tel que "Proportion de Compacteurs par Classe".}

\item{colorsM}{Les couleurs à utiliser pour le graphique des Maisons Mères.}

\item{colorsR}{Les couleurs à utiliser pour le graphique des Régions.}

\item{proportions}{Les données sur la proportion de compacteurs par classe.}

\item{totalTonnageParClasse}{Les données sur le tonnage total par classe.}

\item{totalTonnageParRegion}{Les données sur le tonnage total par région.}

\item{totalTonnageParMaison}{Les données sur le tonnage total par maison mère.}

\item{sitesOutreMer}{Les données sur le nombre de sites en Outre-Mer par classe.}
}
\value{
Un graphique Plotly généré selon le type de vue sélectionné.
}
\description{
BarCy
Crée et affiche des graphiques basés sur le type de vue sélectionné.
}
\examples{
require(dplyr)
r <- NULL
r$data <- Tonnage
r$classe <- stratopt(r$data)

data_copy <- Tonnage
data_copy$classe <- NA
data_copy$classe <- sapply(data_copy$Site, function(site) {
  for(i in 1:5){
    if (site \%in\% r$classe[[paste0("classe",i)]])
    return(i)
  }
  data_copy
})

totalTonnageParClasse <- summarise_data(data_copy,classe,r)

totalTonnageParRegion <- summarise_data(data_copy,Region,r)

totalTonnageParMaison <- summarise_data(data_copy,Maison.Mere,r)


data_copy$IsOutreMer <- CyclamedSampleR::isOutreMer(data_copy)
sites_outre_mer <- summarise_data(data_copy \%>\% filter(IsOutreMer), classe, r)

proportions <- 
data_copy \%>\%
  group_by(classe) \%>\%
  summarise(Count = dplyr::n(),
            ProportionCompacteur = mean(Compacteur),
            MinTonnage = min(Tonnages.DIM, na.rm = TRUE),
            MaxTonnage = max(Tonnages.DIM, na.rm = TRUE)) \%>\%
  mutate(
    class_name = dplyr::case_when(
      classe == 1 ~ paste("\u2264", r$classe$bornes1, " tonnes"),
      classe == 2 ~ paste("[", r$classe$bornes1, ";", r$classe$bornes2, ") tonnes"),
      classe == 3 ~ paste("[", r$classe$bornes2, ";", r$classe$bornes3, ") tonnes"),
      classe == 4 ~ paste("[", r$classe$bornes3, ";", r$classe$bornes4, ") tonnes"),
      classe == 5 ~ paste("\u2265", r$classe$bornes4, " tonnes")
    )
  )

colorsM <- c("red", "green", "blue")
colorsR <- c("orange", "purple", "cyan")

BarCy(viewType = "Tonnage par classe",
      colorsM = colorsM,
      colorsR = colorsR,
      proportions = proportions,
      totalTonnageParClasse = totalTonnageParClasse,
      totalTonnageParRegion = totalTonnageParRegion,
      totalTonnageParMaison = totalTonnageParMaison,
      sitesOutreMer = sites_outre_mer)
}
