% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratopt.R
\name{darkmode}
\alias{darkmode}
\title{Activer le mode Sombre}
\usage{
darkmode(
  session = getDefaultReactiveDomain(),
  bottom = "32px",
  right = "32px",
  left = "unset",
  time = "0.5s",
  mixColor = "#fff",
  backgroundColor = "#fff",
  buttonColorDark = "#100f2c",
  buttonColorLight = "#fff",
  saveInCookies = FALSE,
  label = "🌓",
  autoMatchOsTheme = TRUE
)
}
\arguments{
\item{session}{Session}

\item{bottom}{position basse du boutton}

\item{right}{position droite}

\item{left}{position gauche}

\item{time}{temps de l'animation}

\item{mixColor}{couleur}

\item{backgroundColor}{couleur de fond}

\item{buttonColorDark}{couleur mode sombre}

\item{buttonColorLight}{couleur mode clair}

\item{saveInCookies}{si la valeur est "true", les préférences de l'utilisateur sont enregistrées dans un cookie. La valeur par défaut est false.}

\item{label}{boutton label pour utiliser le mode sombre}

\item{autoMatchOsTheme}{si la valeur est "true", le thème de couleur correspondra automatiquement aux préférences du système d'exploitation de l'utilisateur.}
}
\value{
pas de retour
}
\description{
Permet d'activer le mode sombre avec un boutton
}
\examples{
require(shiny)
vars <- setdiff(names(iris), "Species")
ui <- fluidPage(
  titlePanel("example App"), 
  # Setup
  use_darkmode(),
    sidebarPanel(
      selectInput('xcol', 'X Variable', vars),
     selectInput('ycol', 'Y Variable', vars, selected = vars[[2]]),
     numericInput('clusters', 'Cluster count', 3, min = 1, max = 9)
    )
)
server <- function(input,output,session){
  darkmode(label = '🌓')
  
}
}
\author{
Deepanshu Bhalla
}
