% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{historique}
\alias{historique}
\title{History of random selection of sites}
\format{
## `historique`
A data frame with  rows and 9 columns:
\describe{
  \item{Date}{Date of selection}
  \item{Site1}{site randomly selected from the first class}
  \item{Site2}{site randomly selected from the second class}
  \item{Site3}{site randomly selected from the third class}
  \item{Site4}{site randomly selected from the fourth  class}
  \item{Site5}{site randomly selected from the fifth class}
  \item{Etat}{selection status}
  \item{Commentaire}{comments on the sorting process}
  \item{Annee}{year for which the selection was made}
}
}
\source{
<https://www.cyclamed.org/>
}
\usage{
historique
}
\description{
History of random selection of sites
}
\keyword{datasets}
