% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratopt.R
\name{summarise_data}
\alias{summarise_data}
\title{summarise_data
permet de creer un  data frame tibble selon une variable group by}
\usage{
summarise_data(data, group_var, dataBorne)
}
\arguments{
\item{data}{un dataframe}

\item{group_var}{variable utilisé pour le group by}

\item{dataBorne}{dataframe contenant les bornes}
}
\value{
Un data frame tibble.
}
\description{
summarise_data
permet de creer un  data frame tibble selon une variable group by
}
\examples{
r <- NULL
r$data <- Tonnage
r$classe <- stratopt(r$data)

summarise_data(r$data,Region,r)
}
