% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Tonnage}
\alias{Tonnage}
\title{Data of waste from medicines site information.}
\format{
## `Tonnage`
A data frame with 188 rows and 6 columns:
\describe{
  \item{Region}{Region name}
  \item{Maison.Mere}{Maison.mere name}
  \item{Site}{Site name}
  \item{Tonnages.DIM}{Number of DIM in dumpster (in tons)}
  \item{Nbre.de.rotation}{Number of times dumpster is emptied}
  \item{Compacteur}{1 if the site has a compactor instead of a dumpster, 0 otherwise}
}
}
\source{
<https://www.cyclamed.org/>
}
\usage{
Tonnage
}
\description{
Data of waste from medicines site information.
}
\keyword{datasets}
