% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratopt.R
\name{use_darkmode}
\alias{use_darkmode}
\title{Initialisation du mode sombre}
\usage{
use_darkmode()
}
\value{
pas de retour
}
\description{
Charge les fichiers CSS et JS nécessaires pour le mode sombre.
}
\examples{
require(shiny)
vars <- setdiff(names(iris), "Species")
ui <- fluidPage(
  titlePanel("example App"), 
  # Setup
  use_darkmode(),
    sidebarPanel(
      selectInput('xcol', 'X Variable', vars),
     selectInput('ycol', 'Y Variable', vars, selected = vars[[2]]),
     numericInput('clusters', 'Cluster count', 3, min = 1, max = 9)
    )
)

}
\author{
Deepanshu Bhalla
}
