% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcy.R
\name{BoxCy}
\alias{BoxCy}
\title{BoxCy
permet d'afficher le graphique en boîte du tonnage en fonction des classes}
\usage{
BoxCy(data, selected_region, selected_maison, selectedCompacteur)
}
\arguments{
\item{data}{un dataframe}

\item{selected_region}{Région ou une valeur null}

\item{selected_maison}{Une maison mère une valeur null}

\item{selectedCompacteur}{booleen = "Avec Compacteur"/autre ou une valeur null}
}
\value{
Un graphique en boite du tonnage en fonction des classe
}
\description{
BoxCy
permet d'afficher le graphique en boîte du tonnage en fonction des classes
}
