% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratopt.R
\name{pieCy}
\alias{pieCy}
\title{pieCy
permet de afficher le graphique en secteur de proportion des compacteur}
\usage{
pieCy(proportions_compacteurs)
}
\arguments{
\item{proportions_compacteurs}{un dataframe}
}
\value{
Un graphique pie plot en fonction de la proportion de compacteur
}
\description{
pieCy
permet de afficher le graphique en secteur de proportion des compacteur
}
\examples{
proportions_compacteurs <- 
      Tonnage \%>\%
        mutate(Compacteur = ifelse(Compacteur == 1, "Avec Compacteur", "Sans Compacteur")) \%>\%
        group_by(Compacteur) \%>\%
        summarise(Count = dplyr::n()) \%>\%
        mutate(Proportion = Count / sum(Count))

pieCy(proportions_compacteurs)
}
