% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratopt.R
\name{scatterCy}
\alias{scatterCy}
\title{scatterCy
permet de afficher le graphique de tonnage en fonction des rotations}
\usage{
scatterCy(data, selected_maison, selected_region, selectedCompacteur)
}
\arguments{
\item{data}{un dataframe}

\item{selected_maison}{Une maison mère une valeur null}

\item{selected_region}{Région ou une valeur null}

\item{selectedCompacteur}{booleen = "Avec Compacteur"/autre ou une valeur null}
}
\value{
Un graphique scatter plot en fonction des région/maison mère est la présence d'un compacteur
}
\description{
scatterCy
permet de afficher le graphique de tonnage en fonction des rotations
}
\examples{
scatterCy(Tonnage,NULL,NULL,"Sans Compacteur")
}
